fix test bug and gps protocal calculation

+++ modules/canbus/transport_controller.h
@@ -69,10 +69,10 @@ class TransportController {
   int finishstop_flag = 0;
   int wait_flag = 1;
   int wait_count = 0;
-  int cluopen_last = 0;
-  int brkopen_last = 0;
-  int delta_brk = 0;
-  int delta_clu = 0;
+  double cluopen_last = 0;
+  double brkopen_last = 0;
+  double delta_brk = 0;
+  double delta_clu = 0;
   // control protocol
   CanSender<::apollo::canbus::ChassisDetail> *can_sender_;
   MessageManager<::apollo::canbus::ChassisDetail> *message_manager_;
